+++ src/components/ToxicTextComponent.tsx
