+++ src/components/WizardStateComponent.tsx
